# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Naam Jaap Sewa Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface DevoteeProfile {
  id: string;
  userId: string;
  name: string;
  gotra: string;
  city: string;
  state: string;
  approved: boolean;
  isAdmin: boolean;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  upiId: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}

export interface ProfileEnvelope {
  profile: DevoteeProfile | null;
}

export interface UpsertProfileBody {
  /**
   * @minLength 1
   * @maxLength 80
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 60
   */
  gotra: string;
  /**
   * @minLength 1
   * @maxLength 60
   */
  city: string;
  /**
   * @minLength 1
   * @maxLength 60
   */
  state: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  upiId?: string | null;
  /** @nullable */
  gender?: string | null;
}

export interface Sankalp {
  id: string;
  userId: string;
  date: string;
  accepted: boolean;
  /** @nullable */
  acceptedAt: string | null;
  text: string;
}

export interface SankalpEnvelope {
  sankalp: Sankalp;
  alreadyAcceptedToday: boolean;
}

export interface SankalpHistoryItem {
  date: string;
  accepted: boolean;
  /** @nullable */
  acceptedAt?: string | null;
}

export interface SankalpHistoryEnvelope {
  history: SankalpHistoryItem[];
}

export interface AcceptSankalpBody {
  /** @nullable */
  text?: string | null;
}

export interface PatronSankalpProgress {
  id: string;
  goalCount: number;
  accumulated: number;
  userAccumulated: number;
  remaining: number;
  yajamanaName: string;
  purpose: string;
  ratePerJaap: number;
}

export type JaapSnapshotPendingSankalpAcceptance = {
  [key: string]: unknown;
} | null;

export interface JaapSnapshot {
  todayCount: number;
  totalCount: number;
  todayEarnings: number;
  totalEarnings: number;
  sankalpAccepted: boolean;
  streakDays: number;
  todayMalas: number;
  totalMalas: number;
  activePatronSankalpId?: string | null;
  currentSankalp?: PatronSankalpProgress | null;
  activeMantraText?: string | null;
  allSankalpsDone: boolean;
  pendingSankalpAcceptance?: JaapSnapshotPendingSankalpAcceptance;
}

export interface JaapSnapshotEnvelope {
  snapshot: JaapSnapshot;
}

export interface AddJaapBody {
  /**
   * @minimum 1
   * @maximum 200
   */
  count: number;
  /** @minimum 0 */
  intervalMs: number;
  verified?: boolean;
}

export interface Milestone {
  count: number;
  reward: number;
  achieved: boolean;
  label?: string;
}

export interface DashboardSummary {
  snapshot: JaapSnapshot;
  profile: DevoteeProfile | null;
  milestones: Milestone[];
  dailyTarget: number;
  dailyTargetProgress: number;
}

export interface JaapHistoryDay {
  date: string;
  count: number;
  earnings: number;
}

export interface JaapHistoryEnvelope {
  days: JaapHistoryDay[];
}

export interface LeaderboardEntry {
  rank: number;
  userId?: string;
  name: string;
  gotra: string;
  city: string;
  count: number;
  isMe: boolean;
  badges?: string[];
}

export interface LeaderboardEnvelope {
  scope: string;
  entries: LeaderboardEntry[];
}

export interface GlobalStats {
  totalUsers: number;
  totalJaapAllTime: number;
  totalJaapToday: number;
  totalEarnings: number;
}

export interface AdminUser {
  userId: string;
  profileId: string;
  name: string;
  gotra: string;
  city: string;
  state: string;
  /** @nullable */
  email: string | null;
  approved: boolean;
  isAdmin: boolean;
  totalJaap: number;
  totalEarnings: number;
  suspiciousFlags: number;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  upiId?: string | null;
  createdAt: string;
}

export interface AdminUsersEnvelope {
  users: AdminUser[];
}

export interface AdminUserEnvelope {
  user: AdminUser;
}

export interface SetApprovalBody {
  approved: boolean;
}

export type PayoutStatus = (typeof PayoutStatus)[keyof typeof PayoutStatus];

export const PayoutStatus = {
  pending: "pending",
  paid: "paid",
  rejected: "rejected",
} as const;

export interface Payout {
  id: string;
  userId: string;
  devoteeName: string;
  amount: number;
  status: PayoutStatus;
  /** @nullable */
  upiId?: string | null;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  paymentNote?: string | null;
  requestedAt: string;
  /** @nullable */
  resolvedAt?: string | null;
}

export interface PayoutEnvelope {
  payout: Payout;
}

export interface PayoutsEnvelope {
  payouts: Payout[];
}

export interface PayoutRequest {
  /** @minimum 10 */
  amount: number;
  /** @minLength 3 */
  upiId?: string;
}

export type UpdatePayoutBodyStatus =
  (typeof UpdatePayoutBodyStatus)[keyof typeof UpdatePayoutBodyStatus];

export const UpdatePayoutBodyStatus = {
  paid: "paid",
  rejected: "rejected",
} as const;

export type UpdatePayoutBodyPaymentMethod =
  (typeof UpdatePayoutBodyPaymentMethod)[keyof typeof UpdatePayoutBodyPaymentMethod];

export const UpdatePayoutBodyPaymentMethod = {
  upi: "upi",
  bank_transfer: "bank_transfer",
  cash: "cash",
} as const;

export interface UpdatePayoutBody {
  status: UpdatePayoutBodyStatus;
  paymentMethod?: UpdatePayoutBodyPaymentMethod;
  paymentNote?: string;
}

export interface AdminStats {
  totalUsers: number;
  approvedUsers: number;
  pendingApproval: number;
  totalJaap: number;
  totalPayoutPending: number;
  totalPayoutPaid: number;
  todayJaap: number;
  todayExpense: number;
  totalExpense: number;
  activeBhaktsToday: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type GetJaapHistoryParams = {
  /**
   * @minimum 1
   * @maximum 90
   */
  days?: number;
};

export type GetLeaderboardParams = {
  scope?: GetLeaderboardScope;
};

export type GetLeaderboardScope =
  (typeof GetLeaderboardScope)[keyof typeof GetLeaderboardScope];

export const GetLeaderboardScope = {
  today: "today",
  week: "week",
  alltime: "alltime",
} as const;